post detail and edit post moved to Posts folder
