+++ src/pages/Posts/EditPost.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
-import Spinner from "../components/UI/Spinner";
-import PostModel from "../models/post";
+import Spinner from "../../components/UI/Spinner";
+import PostModel from "../../models/post";
 import "./EditPost.css";
 
 const EditPost = (props) => {
